requirements.txt
apipkg==1.4
appdirs==1.4.0
astroid==1.4.7
bandit==0.17.3
block-disposable-email==1.0.1
coverage==4.1
dj-database-url==0.4.1
Django==1.8.7
django-disposable-email-checker==1.2.1
django-extensions==1.6.1
djangorestframework==3.3.2
djangorestframework-jwt==1.7.2
django_seed
docopt==0.6.2
dodgy==0.1.9
execnet==1.4.1
factory-boy==2.6.0
fake-factory==0.5.9
gitdb==0.6.4
GitPython==2.0.7
gunicorn==19.6.0
isort==4.2.5
jedi==0.9.0
lazy-object-proxy==1.2.2
logilab-common==1.2.2
mccabe==0.5.0
mock==1.3.0
newrelic
pbr==1.10.0
pep257==0.7.0
pep8==1.7.0
pep8-naming==0.4.1
prompt-toolkit==0.57
prospector==0.11.7
psycopg2==2.6.1
ptpython==0.28
py==1.4.31
pyflakes==1.2.3
Pygments==2.1.3
PyJWT==1.4.0
pylint==1.6.3
pylint-celery==0.3
pylint-common==0.2.2
pylint-django==0.7.2
pylint-flask==0.3
pylint-plugin-utils==0.2.3
pytest==2.8.5
pytest-cov==2.2.0
pytest-django==2.9.1
pytest-pythonpath==0.7
pytest-xdist==1.14
python-dateutil==2.4.2
PyYAML==3.11
raven==5.8.1
requirements-detector==0.5.1
setoptconf==0.2.0
six==1.10.0
smmap==0.9.0
stevedore==1.16.0
wcwidth==0.1.7
whitenoise==2.0.6
wrapt==1.10.8
